requirements.txt
surrealdb
discordpy
python-dotenv
requests